requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
deepface==0.0.75
tensorflow==2.13.0
Pillow==10.0.0
requests==2.31.0
werkzeug==3.0.0